# lib/api-zod/src/generated/types/scanInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SentinelX API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanInputType } from './scanInputType';

export interface ScanInput {
  /** @minLength 1 */
  name: string;
  type: ScanInputType;
}